Efficiency internal.closestIntensities
Improved efficiency by removing internal.calibCombine from function and having an input parameter as calibration_df. This should not be touched by a user, so shouldn't need as many guardrails as an external facing function. It is also intended to be called by higher level functions, which will have already done the internal.calibCombine step